% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_deg.R
\name{collapse.deg}
\alias{collapse.deg}
\title{A function to filter out probes standing for the same gene which have
contradictory DE direction}
\usage{
collapse.deg(geo2r)
}
\arguments{
\item{geo2r}{data.frame/data.table output of the deg.def() function}
}
\description{
This function check if the probes standing for the same gene has a unique 
value for the variable deg from deg.def()
}
\examples{
collapse.deg()
}
\keyword{Probe}
\keyword{collapse}
