% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_mv.R
\name{combining_mv}
\alias{combining_mv}
\title{A function to draw the 'Combining meta-analysis' MetaVolcano}
\usage{
combining_mv(
  diffexp = list(),
  pcriteria = "pvalue",
  foldchangecol = "Log2FC",
  genenamecol = "Symbol",
  geneidcol = NULL,
  metafc = "Mean",
  metathr = 0.01,
  collaps = "FALSE",
  jobname = "MetaVolcano",
  outputfolder = tempdir(),
  draw = "HTML",
  colors = c("#083e46", "grey", "#811820"),
  point_size = 0.5,
  label_genes = NULL,
  label_top_n = NULL,
  label_size = 3,
  plot_title = NULL,
  show_legend = FALSE
)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{pcriteria}{the column name of the Pval criteria to consider 
c("adj.P.Val", "P.Value") <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript
variable <string>}

\item{metafc}{method for summarizing gene fold-changes across studies 
c("Mean", "Median") <string>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}

\item{collaps}{if probes should be collapsed based on the DE direction
<logical>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw the .pdf or .html visualization 
<c(NULL, "PDF", "HTML")>}

\item{colors}{vector of colors for the plot c(down, neutral, up)}

\item{point_size}{size of points in the plot}

\item{label_genes}{character vector of specific genes to label}

\item{label_top_n}{number of top genes to label automatically}

\item{label_size}{size of gene name labels}

\item{plot_title}{custom plot title}

\item{show_legend}{whether to show the legend}
}
\value{
\code{MetaVolcano} object
}
\description{
This function draws the 'Combining meta-analysis' MetaVolcano
}
\examples{
data(diffexplist)
mv <- combining_mv(diffexplist)
str(mv)
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
