% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_freq_data.R
\name{cum.freq.data}
\alias{cum.freq.data}
\title{A data formating function for inverse-cummulative DEG distribution}
\usage{
cum.freq.data(meta_geo2r, nstud)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\description{
This function counts how many genes consistly appears as DE along the inputed GEO2R outputs
}
\examples{
cum.freq.data()
}
\keyword{DEG}
\keyword{data}
\keyword{distribution}
\keyword{format}
\keyword{inverse-cummulative}
