% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_freq_data.R
\name{cum_freq_data}
\alias{cum_freq_data}
\title{A data formating function for inverse-cummulative DEG distribution}
\usage{
cum_freq_data(meta_diffexp, nstud)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing all the input studies}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\description{
This function counts how many genes consistly appears as DE along the 
input studies
}
\examples{
cum_freq_data()
}
\keyword{DEG}
\keyword{data}
\keyword{distribution}
\keyword{format}
\keyword{inverse-cummulative}
