% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_deg.R
\name{deg.def}
\alias{deg.def}
\title{A DEG definition function}
\usage{
deg.def(geo2r, columnStatistics, columnlogFC, pvalue, logfc)
}
\arguments{
\item{geo2r}{data.frame/data.table with DE results where lines are genes}

\item{columnStatistics}{column name of the pvalue variable <strings>}

\item{columnlogFC}{column name of the foldchange variable <string>}

\item{pvalue}{thresholds <ex: 0.05>}

\item{logfc}{foldchange threshold <ex: 0.4>}
}
\description{
This function creates a new variable indicating DEGs as -1, 0, 1 based on the
user-defined foldchange and p-value criteria
}
\examples{
deg.def()
}
\keyword{DEG}
\keyword{definition}
