% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_deg.R
\name{deg_def}
\alias{deg_def}
\title{A DEG definition function}
\usage{
deg_def(diffexp, pcriteria, foldchangecol, pv, fc)
}
\arguments{
\item{diffexp}{data.frame/data.table with differential expression results}

\item{pcriteria}{column name of the pvalue variable <strings>}

\item{foldchangecol}{column name of the foldchange variable <string>}

\item{pv}{pvalue threshold <double>}

\item{fc}{foldchange threshold <double>}
}
\value{
\code{data.table/data.frame} with a new deg variable
}
\description{
This function creates a new variable indicating DEGs as -1, 0, 1 based on the
user-defined foldchange and p-value criteria
}
\examples{
deg_def()
}
\keyword{DEG}
\keyword{definition}
