% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffexplist}
\alias{diffexplist}
\title{Differential expression results from five studies}
\format{
A named list with 5 data frames with ~20k genes and 5 variables:
\describe{
  \item{GSE12050}{differential expression result, disease vs healthy}
  \item{GSE24883}{differential expression result, disease vs healthy}
  ...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
diffexplist
}
\description{
A named list with five differential expresion results.
}
\keyword{datasets}
