% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_metafor.R
\name{do.metafor}
\alias{do.metafor}
\title{A function to run the 'Random Effect Model' MetaVolcano}
\usage{
\method{do}{metafor}(geo2r_res, pcriteria, genenamecol, geneidcol,
  foldchangecol, pvalue, logfc, collaps, jobname, outputfolder, ncores,
  cvar)
}
\arguments{
\item{geo2r_res}{list of data.frame/data.table (s) with DE results where lines are genes}

\item{pcriteria}{the column name of the Pval criteria to consider c("adj.P.Val", "P.Value") <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript variable <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{pvalue}{the Pval to use as threshold c(0:1) <double>}

\item{logfc}{the foldchange to use as DE threshold c(-Inf: Inf) <double>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{ncores}{the number of processors the user wants to use <integer>}

\item{cvar}{weather or not to calculate gene variance from confidence interval limits <logical>}
}
\description{
This function runs the 'Random Effect Model' MetaVolcano section
}
\examples{
do.metafor()
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
