% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cum_freq.R
\name{draw.cum.freq}
\alias{draw.cum.freq}
\title{A function to visualize the inverse-cummulative DEG distribution}
\usage{
draw.cum.freq(meta_geo2r, nstud)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\description{
This function create a ggplotly object with the inverse-cummulative DEG distribution
}
\examples{
draw.cum.freq()
}
\keyword{DEG}
\keyword{distribution}
\keyword{draw}
\keyword{inverse-cummulative}
