% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_degbar_cum.R
\name{draw.degbar.cum}
\alias{draw.degbar.cum}
\title{A function for the initial MetaVolcano overview}
\usage{
draw.degbar.cum(geo2r_res, pcriteria, foldchangecol, genenamecol,
  geneidcol, pvalue, logfc, collaps, jobname, outputfolder, draw, ncores)
}
\arguments{
\item{geo2r_res}{list of data.frame/data.table (s) with DE results where lines are genes}

\item{pcriteria}{the column name of the Pval criteria to consider <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript variable <string>}

\item{pvalue}{the Pval to use as threshold c(0:1) <double>}

\item{logfc}{the foldchange to use as DE threshold c(-Inf: Inf) <double>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw the .html visualization}

\item{ncores}{the number of processors the user wants to use <integer>}
}
\description{
This function runs the first Meta-Volcano section 
i) DEGs per study. Write .html barplot interactive plotly 
ii) Inverse-cummulative DEG distribution along the inputed GEO2R outputs. Write .html lineplot interactive plotly 
iii) Merge the inputed GEO2R outputs in one data.frame/data.table. Output the meta_geo2r data.frame object
}
\examples{
draw.degbar.cum()
}
\keyword{metavolcano}
\keyword{write}
