% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_degbar.R
\name{draw.degbar}
\alias{draw.degbar}
\title{A function for DEG barplot visualization}
\usage{
draw.degbar(degbar_data)
}
\arguments{
\item{degbar_data}{output of the set.degbar.data() function <data.fram/data.table>}
}
\description{
This function visualize as barplots the number of DEGs across the inputed GEO2R outputs
}
\examples{
draw.degbar()
}
\keyword{DEG}
\keyword{barplot}
\keyword{draw}
