% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_forest.R
\name{draw.forest}
\alias{draw.forest}
\title{A function to draw a forest plot from the REM MetaVolcano result}
\usage{
draw.forest(gene, genecol, remres, foldchangecol, llcol, rlcol, studynames,
  jobname, outputfolder, draw)
}
\arguments{
\item{gene}{query gene to plot}

\item{genecol}{name of the variable with genes <string>}

\item{remres}{data.table/data.frame output of the do.metafor function <data.table/data.frame>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{llcol}{left limit of the fold change coinfidence interval variable name <string>}

\item{rlcol}{right limit of the fold change coinfidence interval variable name <string>}

\item{studynames}{names of the input studies <character vector>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/ <string>}

\item{draw}{either 'PDF' or 'HTML' to save metaolcano as .pdf or .html respectively <string>}
}
\description{
This function draws a forest plot for a given gene based on the REM MetaVolcano result
}
\examples{
draw.forest()
}
\keyword{a}
\keyword{draw}
\keyword{for}
\keyword{forest}
\keyword{gene}
\keyword{given}
\keyword{plot}
