% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_metavolcano_metap.R
\name{draw.metavolcano.metap}
\alias{draw.metavolcano.metap}
\title{A function to draw the 'Combining meta-analysis' MetaVolcano}
\usage{
draw.metavolcano.metap(meta_geo2r, pcriteria, metathr, nstud, jobname,
  collaps, metap, metafc, outputfolder)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{pcriteria}{the column name of the Pval criteria to consider c("adj.P.Val", "P.Value") <string>}

\item{metathr}{the proportion of top meta-DEGs to be highlighted 'combining meta-analysis'<double>}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}

\item{jobname}{name of the running job <string>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{metap}{method for Pval combination/merge. c("Fisher")  <string>}

\item{metafc}{method for summarizing gene fold-changes across studies c("Mean", "Median") <string>}

\item{outputfolder}{/path where to write the results/ <string>}
}
\description{
This function draws the 'Combining meta-analysis' MetaVolcano
}
\examples{
draw.metavolcano.metap()
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
