% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mv_gplotly.R
\name{draw.mv.gplotly}
\alias{draw.mv.gplotly}
\title{A MetaVolcano drawing function}
\usage{
draw.mv.gplotly(meta_geo2r, nstud, metathr, genecol, metap)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}

\item{genecol}{column name of the variable to label genes in the .html file <string>}

\item{metap}{wheather or not the drawing is for the combining-metavolcano  <logical>}
}
\description{
This function creates a ggplotly object with the MetaVolcano
}
\examples{
draw.mv.gplotly()
}
\keyword{draw}
\keyword{metavolcano}
