% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cum_freq.R
\name{draw_cum_freq}
\alias{draw_cum_freq}
\title{A function to visualize the inverse-cummulative DEG distribution}
\usage{
draw_cum_freq(meta_diffexp, nstud)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing all the input studies}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\description{
This function create a ggplot object with the inverse-cummulative
DEG distribution
}
\examples{
draw_cum_freq()
}
\keyword{DEG}
\keyword{distribution}
\keyword{draw}
\keyword{inverse-cummulative}
