% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_degbar.R
\name{draw_degbar}
\alias{draw_degbar}
\title{A function for DEG barplot visualization}
\usage{
draw_degbar(degbar_data)
}
\arguments{
\item{degbar_data}{output of the \code{set_degbar_data()} function 
<data.fram/data.table>}
}
\value{
\code{ggplot2} object
}
\description{
This function visualize as barplots the number of DEGs across the input 
studies
}
\examples{
data(diffexplist)
diffexp <- lapply(diffexplist, function(...) deg_def(..., "pvalue", 
           "Log2FC", 0.05, 0))
bardat <- set_degbar_data(diffexp)
gg <- draw_degbar(bardat)
plot(gg)
}
\keyword{DEG}
\keyword{barplot}
\keyword{draw}
