% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_featurebar.R
\name{draw_featurebar}
\alias{draw_featurebar}
\title{A function for feature barplot visualization}
\usage{
draw_featurebar(degbar_data, colors = c("#811820", "grey90", "#083e46"))
}
\arguments{
\item{degbar_data}{output of the \code{set_degbar_data()} function 
<data.fram/data.table>}
}
\value{
\code{ggplot2} object
}
\description{
This function visualize as barplots the number of features across the input 
studies
}
\examples{
data(diffexplist)
diffexp <- lapply(diffexplist, function(...) deg_def(..., "pvalue", 
           "Log2FC", 0.05, 0))
bardat <- set_degbar_data(diffexp)
gg <- draw_featurebar(bardat)
plot(gg)
}
\keyword{barplot}
\keyword{draw}
\keyword{feature}
