% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_geo2rformat.R
\name{if.geo2rformat}
\alias{if.geo2rformat}
\title{Checking if input files have GEO2R format}
\usage{
if.geo2rformat(geo2r_res_file, inputfolder)
}
\arguments{
\item{geo2r_res_file}{<GEO2R file>}

\item{inputfolder}{</input folder/>}
}
\description{
Logical value
}
\examples{
if.geo2rformat()
}
\keyword{input}
\keyword{validation}
