% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_mv.R
\docType{class}
\name{MetaVolcano-class}
\alias{MetaVolcano-class}
\title{An S4 class to represent MetaVolcanoR results}
\description{
An S4 class to represent MetaVolcanoR results
}
\section{Slots}{

\describe{
\item{\code{input}}{merged diiferential expression inputs \code{data.frame}}

\item{\code{inputnames}}{names of the differential expression inputs \code{character}}

\item{\code{metaresult}}{meta-analysis results \code{data.frame}}

\item{\code{MetaVolcano}}{plot with meta-analysis results}

\item{\code{degfreq}}{supplementary figure of the vote-counting MetaVolcano}
}}

