% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mv.R
\name{plot_mv}
\alias{plot_mv}
\title{A MetaVolcano ploting function}
\usage{
plot_mv(
  meta_diffexp,
  nstud,
  genecol,
  comb,
  metafc,
  colors = c("#083e46", "grey", "#811820"),
  point_size = 0.5,
  label_genes = NULL,
  label_top_n = NULL,
  label_size = 3,
  plot_title = NULL,
  show_legend = FALSE
)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing the differential 
expression inputs}

\item{nstud}{the number of differential expression inputs <integer>}

\item{genecol}{column name of the variable to label genes in the .html 
file <string>}

\item{comb}{wheather or not the drawing is for the combining-metavolcano 
<logical>}

\item{metafc}{method for summarizing gene fold-changes across studies 
c("Mean", "Median") <string>}

\item{colors}{vector of colors for the plot}

\item{point_size}{size of the points}

\item{label_genes}{character vector of specific genes to label}

\item{label_top_n}{number of top genes to label}

\item{label_size}{size of gene labels}

\item{plot_title}{custom plot title}

\item{show_legend}{whether to show the legend}
}
\value{
\code{ggplot2} object
}
\description{
This function plots either the combining- or the vote-counting- MetaVolcanos
}
\examples{
data(diffexplist)
mv <- votecount_mv(diffexplist)
gg <- plot_mv(mv@metaresult, length(diffexplist), "Symbol", FALSE, "Mean")
plot(gg)
}
\keyword{draw}
\keyword{metavolcano}
