% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename.col}
\alias{rename.col}
\title{A column renaming function for GEO2R outputs}
\usage{
rename.col(geo2r_list, genenamecol, geneidcol, collaps, ncores)
}
\arguments{
\item{geo2r_list}{list of data.frame/data.table (s) with DE results where lines are genes}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript variable <string>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{ncores}{the number of processors the user wants to use  <integer>}
}
\description{
This function rename the probe and gene name columns from the GEO2R outputs
}
\examples{
rename.col()
}
\keyword{column}
\keyword{rename}
