% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmodel.R
\name{rnmodel}
\alias{rnmodel}
\title{A function to model foldchange variance along several studies}
\usage{
rnmodel(gene)
}
\arguments{
\item{gene}{named vector with foldchanges and variances (lines of the meta_geo2r: data.frame/data.table containing all the inputed GEO2R outputs)}
}
\description{
This function calculate a Random Effect Model summary
}
\examples{
rnmodel()
}
\keyword{REM}
\keyword{summary}
