% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_degbar_data.R
\name{set.degbar.data}
\alias{set.degbar.data}
\title{A function setting data format for DEG barplot visualization}
\usage{
set.degbar.data(geo2r_res)
}
\arguments{
\item{geo2r_res}{list of data.frame/data.table (s) output of the deg.def() function <list>}
}
\description{
This function summarize the variable deg from the deg.def() function to visualize as barplots the number of DEGs across the inputed GEO2R outputs
}
\examples{
set.degbar.data()
}
\keyword{DEG}
\keyword{barplot}
\keyword{data}
\keyword{format}
\keyword{set}
