% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_degbar_data.R
\name{set_degbar_data}
\alias{set_degbar_data}
\title{A function setting data format for DEG barplot visualization}
\usage{
set_degbar_data(diffexp)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) output of the deg_def() function <list>}
}
\description{
This function summarize the variable deg from the deg_def() function to visualize as barplots the number of DEGs per inputed study
}
\examples{
set_degbar_data()
}
\keyword{DEG}
\keyword{barplot}
\keyword{data}
\keyword{format}
\keyword{set}
