% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_mv.R
\name{combining_mv}
\alias{combining_mv}
\title{A function to draw the 'Combining meta-analysis' MetaVolcano}
\usage{
combining_mv(diffexp = list(), pcriteria = "pvalue",
  foldchangecol = "Log2FC", genenamecol = "Symbol", geneidcol = NULL,
  metafc = "Mean", metathr = 0.01, collaps = "FALSE",
  jobname = "MetaVolcano", outputfolder = ".", draw = "HTML",
  ncores = 1)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{pcriteria}{the column name of the Pval criteria to consider 
c("adj.P.Val", "P.Value") <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript
variable <string>}

\item{metafc}{method for summarizing gene fold-changes across studies 
c("Mean", "Median") <string>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}

\item{collaps}{if probes should be collapsed based on the DE direction
<logical>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw the .pdf or .html visualization 
<c(NULL, "PDF", "HTML")>}

\item{ncores}{the number of processors the user wants to use <integer>}
}
\value{
\code{MetaVolcano} object
}
\description{
This function draws the 'Combining meta-analysis' MetaVolcano
}
\examples{
data(diffexplist)
mv <- combining_mv(diffexplist)
str(mv)
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
