% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_freq_data.R
\name{cum_freq_data}
\alias{cum_freq_data}
\title{A data formating function for inverse-cummulative DEG distribution}
\usage{
cum_freq_data(meta_diffexp, nstud)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing all the input studies}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\value{
\code{data.frame} inverse cummulative distribution
}
\description{
This function counts how many genes consistly appears as DE along the 
input studies
}
\examples{
data(diffexplist)
diffexp <- lapply(diffexplist, function(...) deg_def(..., "pvalue", 
           "Log2FC", 0.05, 0))
diffexp <- rename_col(diffexp, "Symbol")
meta_diffexp <- Reduce(function(...) merge(..., by = "Symbol", all = TRUE),
           diffexp)
meta_diffexp \%>\%
dplyr::select(dplyr::matches("deg_")) \%>\%
    data.matrix -> n_deg
meta_diffexp[['ndeg']] <- rowSums(n_deg^2, na.rm = TRUE)
cfd <- cum_freq_data(meta_diffexp, length(diffexplist))
head(cfd, 3)
}
\keyword{DEG}
\keyword{data}
\keyword{distribution}
\keyword{format}
\keyword{inverse-cummulative}
