% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_deg.R
\name{deg.def}
\alias{deg.def}
\title{A DEG definition function}
\usage{
deg.def(diffexp, pcriteria, foldchangecol, pv, fc)
}
\arguments{
\item{diffexp}{data.frame/data.table with DE results where lines are genes}

\item{pcriteria}{column name of the pvalue variable <strings>}

\item{foldchangecol}{column name of the foldchange variable <string>}

\item{pv}{thresholds <ex: 0.05>}

\item{fc}{foldchange threshold <ex: 0.4>}
}
\description{
This function creates a new variable indicating DEGs as -1, 0, 1 based on the
user-defined foldchange and p-value criteria
}
\examples{
deg.def()
}
\keyword{DEG}
\keyword{definition}
