% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_metav.R
\name{draw.metav}
\alias{draw.metav}
\title{A function to draw the 'Random Effect Model' MetaVolcano}
\usage{
draw.metav(meta_res, jobname, outputfolder, genecol)
}
\arguments{
\item{meta_res}{data.frame/data.table containing all the inputed GEO2R outputs, and after REM calculation}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/ <string>}

\item{genecol}{column name of the variable to label genes in the .html file <string>}
}
\description{
This function draws the 'Random Effect Model' MetaVolcano
}
\examples{
draw.metavolcano.metap()
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
