% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_metavolcano.R
\name{draw.metavolcano}
\alias{draw.metavolcano}
\title{A function to draw the 'Vote-counting meta-analysis' MetaVolcano}
\usage{
draw.metavolcano(meta_geo2r, metathr, nstud, jobname, collaps,
  outputfolder)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{metathr}{the proportion of studies where a gene has to be consistently DE to be highlighted 'vote-counting meta-analysis'}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}

\item{jobname}{name of the running job <string>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{outputfolder}{/path where to write the results/ <string>}
}
\description{
This function draws the 'Vote-counting meta-analysis' MetaVolcano
}
\examples{
draw.metavolcano()
}
\keyword{'vote-counting}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
