% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mv_gplotly.R
\name{draw.mv.gplotly}
\alias{draw.mv.gplotly}
\title{A MetaVolcano drawing function}
\usage{
draw.mv.gplotly(meta_geo2r, nstud, metathr, collaps, metap)
}
\arguments{
\item{meta_geo2r}{data.frame/data.table containing all the inputed GEO2R outputs}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}

\item{metathr}{the proportion of studies where a gene has to be consistently DE to be highlighted 'vote-counting meta-analysis' or the proportion of top meta-DEGs to be highlighted 'combining meta-analysis'<double>}

\item{collaps}{if probes should be collapsed based on the DE direction <logical>}

\item{metap}{if the 'combining meta-analysis' strategy should be used <logical>}
}
\description{
This function creates a ggplotly object with the MetaVolcano
}
\examples{
draw.mv.gplotly()
}
\keyword{draw}
\keyword{metavolcano}
