% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cum_freq.R
\name{draw_cum_freq}
\alias{draw_cum_freq}
\title{A function to visualize the inverse-cummulative DEG distribution}
\usage{
draw_cum_freq(meta_diffexp, nstud)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing all the input studies}

\item{nstud}{the number of inputed GEO2R outputs  <integer>}
}
\value{
\code{ggplot2} object
}
\description{
This function create a ggplot object with the inverse-cummulative
DEG distribution
}
\examples{
data(diffexplist)
diffexp <- lapply(diffexplist, function(...) deg_def(..., "pvalue", 
           "Log2FC", 0.05, 0))
diffexp <- rename_col(diffexp, "Symbol", 1)
meta_diffexp <- Reduce(function(...) merge(..., by = "Symbol", all = TRUE),
           diffexp)
gg <- draw_cum_freq(meta_diffexp, length(diffexplist))
plot(gg)
}
\keyword{DEG}
\keyword{distribution}
\keyword{draw}
\keyword{inverse-cummulative}
