% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_degbar.R
\name{draw_degbar}
\alias{draw_degbar}
\title{A function for DEG barplot visualization}
\usage{
draw_degbar(degbar_data)
}
\arguments{
\item{degbar_data}{output of the \code{set_degbar_data()} function 
<data.fram/data.table>}
}
\description{
This function visualize as barplots the number of DEGs across the input 
studies
}
\examples{
draw_degbar()
}
\keyword{DEG}
\keyword{barplot}
\keyword{draw}
