% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_forest.R
\name{draw_forest}
\alias{draw_forest}
\title{A function to draw a forest plot from the REM MetaVolcano result}
\usage{
draw_forest(remres, gene = "A2M", genecol = "Symbol",
  foldchangecol = "Log2FC", llcol = "CI.L", rlcol = "CI.R",
  jobname = "MetaVolcano", outputfolder = ".", draw = "PDF")
}
\arguments{
\item{remres}{MetaVolcano object. Output of the rem_mv() function
<MetaVolcano>}

\item{gene}{query gene to plot}

\item{genecol}{name of the variable with genes <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{llcol}{left limit of the fold change coinfidence interval variable
name <string>}

\item{rlcol}{right limit of the fold change coinfidence interval variable
name <string>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/ <string>}

\item{draw}{either 'PDF' or 'HTML' to save metaolcano as .pdf or .html
respectively <string>}
}
\value{
\code{ggplot2} object
}
\description{
This function draws a forest plot for a given gene based on the REM 
MetaVolcano result
}
\examples{
draw_forest()
}
\keyword{a}
\keyword{draw}
\keyword{for}
\keyword{forest}
\keyword{gene}
\keyword{given}
\keyword{plot}
