% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_forest.R
\name{draw_forest}
\alias{draw_forest}
\title{A function to draw a forest plot from the REM MetaVolcano result}
\usage{
draw_forest(
  remres,
  gene = "MMP9",
  genecol = "Symbol",
  foldchangecol = "Log2FC",
  llcol = "CI.L",
  rlcol = "CI.R",
  jobname = "MetaVolcano",
  outputfolder = ".",
  draw = "PDF"
)
}
\arguments{
\item{remres}{MetaVolcano object. Output of the rem_mv() function
<MetaVolcano>}

\item{gene}{query gene to plot}

\item{genecol}{name of the variable with genes <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{llcol}{left limit of the fold change coinfidence interval variable
name <string>}

\item{rlcol}{right limit of the fold change coinfidence interval variable
name <string>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/ <string>}

\item{draw}{either 'PDF' or 'HTML' to save metaolcano as .pdf or .html
respectively <string>}
}
\value{
\code{ggplot2} object
}
\description{
This function draws a forest plot for a given gene based on the REM 
MetaVolcano result
}
\examples{
data(diffexplist)
diffexplist <- lapply(diffexplist, function(del) {
    dplyr::filter(del, grepl("MP", Symbol))
})
mv <- rem_mv(diffexplist, metathr = 0.1)
gg <- draw_forest(mv, gene="MMP9")
plot(gg)
}
\keyword{draw}
\keyword{forest-plot}
\keyword{gene}
