% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MetaVolcano-class}
\alias{MetaVolcano-class}
\title{MetaVolcano S4 Class}
\description{
MetaVolcano S4 Class
}
\section{Slots}{

\describe{
\item{\code{input}}{data.frame}

\item{\code{inputnames}}{character}

\item{\code{metaresult}}{data.frame}

\item{\code{MetaVolcano}}{gg}

\item{\code{degfreq}}{gg}
}}

