% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mv.R
\name{plot_mv}
\alias{plot_mv}
\title{A MetaVolcano ploting function}
\usage{
plot_mv(meta_diffexp, nstud, genecol, comb, metafc)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing the differential 
expression inputs}

\item{nstud}{the number of differential expression inputs <integer>}

\item{genecol}{column name of the variable to label genes in the .html 
file <string>}

\item{comb}{wheather or not the drawing is for the combining-metavolcano 
<logical>}

\item{metafc}{method for summarizing gene fold-changes across studies 
c("Mean", "Median") <string>}
}
\value{
\code{ggplot2} object
}
\description{
This function plots either the combining- or the vote-counting- MetaVolcanos
}
\examples{
plot_mv()
}
\keyword{draw}
\keyword{metavolcano}
