% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rem.R
\name{plot_rem}
\alias{plot_rem}
\title{A function to plot the Random Effect Model (REM) MetaVolcano}
\usage{
plot_rem(meta_diffexp, jobname, outputfolder, genecol, metathr)
}
\arguments{
\item{meta_diffexp}{data.frame/data.table containing the REM results from 
rem_mv() <data.table/data.frame>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/ <string>}

\item{genecol}{column name of the variable to label genes in the .html file
<string>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}
}
\value{
\code{ggplot2} object
}
\description{
This function plots the REM MetaVolcano using ggplot2
}
\examples{
data(diffexplist)
diffexplist <- lapply(diffexplist, function(del) {
    dplyr::filter(del, grepl("MP", Symbol))
})
mv <- rem_mv(diffexplist, metathr = 0.1)
gg <- plot_rem(mv@metaresult, "MV", ".", "Symbol", 0.01)
plot(gg)
}
\keyword{REM}
\keyword{metavolcano}
\keyword{write}
