% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_mv.R
\name{rem_mv}
\alias{rem_mv}
\title{A function to perform the Random Effect Model (REM) MetaVolcano}
\usage{
rem_mv(diffexp = list(), pcriteria = "pvalue",
  foldchangecol = "Log2FC", genenamecol = "Symbol", geneidcol = NULL,
  collaps = FALSE, llcol = "CI.L", rlcol = "CI.R", vcol = NULL,
  cvar = TRUE, metathr = 0.01, jobname = "MetaVolcano",
  outputfolder = ".", draw = "HTML", ncores = 1)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{pcriteria}{the column name of the pvalue variable <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript
variable <string>}

\item{collaps}{if probes should be collapsed based on the DE direction
<logical>}

\item{llcol}{left limit of the fold change coinfidence interval variable
name <string>}

\item{rlcol}{right limit of the fold change coinfidence interval variable
name <string>}

\item{vcol}{name of the fold change variance variable <string>}

\item{cvar}{weather or not to calculate gene variance from confidence 
interval limits <logical>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw the .html visualization <logical>}

\item{ncores}{the number of processors the user wants to use <integer>}
}
\value{
MetaVolcano object
}
\description{
This function runs the 'Random Effect Model' MetaVolcano section
}
\examples{
data(diffexplist)
diffexplist <- lapply(diffexplist, function(del) {
    dplyr::filter(del, grepl("MP", Symbol))
})
mv <- rem_mv(diffexplist, metathr = 0.1)
str(mv)
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
