% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_mv.R
\name{rem_mv}
\alias{rem_mv}
\title{Random Effect Model (REM) MetaVolcano}
\usage{
rem_mv(
  diffexp = list(),
  pcriteria = "pvalue",
  foldchangecol = "Log2FC",
  genenamecol = "Symbol",
  geneidcol = NULL,
  collaps = FALSE,
  llcol = "CI.L",
  rlcol = "CI.R",
  vcol = NULL,
  cvar = TRUE,
  metathr = 0.01,
  jobname = "MetaVolcano",
  outputfolder = tempdir(),
  draw = "HTML",
  ncores = 1,
  colors = c(low = "blue", mid = "white", high = "red", na = "grey80"),
  point_size = 0.6,
  label_genes = NULL,
  label_top_n = NULL,
  label_size = 3,
  plot_title = NULL,
  show_legend = TRUE
)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{pcriteria}{the column name of the pvalue variable <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript
variable <string>}

\item{collaps}{if probes should be collapsed based on the DE direction
<logical>}

\item{llcol}{left limit of the fold change coinfidence interval variable
name <string>}

\item{rlcol}{right limit of the fold change coinfidence interval variable
name <string>}

\item{vcol}{name of the fold change variance variable <string>}

\item{cvar}{weather or not to calculate gene variance from confidence
interval limits <logical>}

\item{metathr}{top percentage of perturbed genes to be highlighted <double>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw the .html visualization <logical>}

\item{ncores}{the number of processors the user wants to use <integer>}

\item{colors}{named vector of colors for gradient: c(low, mid, high, na)}

\item{point_size}{size of points in the plot}

\item{label_genes}{character vector of specific genes to label}

\item{label_top_n}{number of top genes (by p-value) to label automatically}

\item{label_size}{size of gene name labels}

\item{plot_title}{custom plot title (NULL for no title)}

\item{show_legend}{whether to show the legend (TRUE/FALSE)}
}
\value{
MetaVolcano object
}
\description{
This function performs Random Effect Model meta-analysis on differential expression results.
This function runs the 'Random Effect Model' MetaVolcano section
}
\examples{
data(diffexplist)
diffexplist <- lapply(diffexplist, function(del) {
    dplyr::filter(del, grepl("MP", Symbol))
})
}
\keyword{'combining}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
