% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remodel.R
\name{remodel}
\alias{remodel}
\title{A function to model foldchange variance along several studies  
This function calculate the REM-summary fold-change}
\usage{
remodel(gene, foldchangecol, vcol)
}
\arguments{
\item{gene}{named vector with foldchanges and variances <vector<}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{vcol}{name of the fold change variance variable <string>}
}
\value{
\code{data.frame} with REM results for a gene
}
\description{
A function to model foldchange variance along several studies  
This function calculate the REM-summary fold-change
}
\examples{
remodel()
}
\keyword{REM}
\keyword{summary}
