% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\title{A column renaming function merged inputs}
\usage{
rename_col(diffexp, genecol, ncores)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{genecol}{the column name of the geneID or gene name variable <string>}

\item{ncores}{the number of processors the user wants to use  <integer>}
}
\value{
\code{data.tabledata.frame} with new colnames
}
\description{
This function rename the columns of the merged inputs
}
\examples{
data(diffexplist)
lapply(diffexplist, colnames)
diffexp <- rename_col(diffexplist, "Symbol", 1)
lapply(diffexp, colnames)
}
\keyword{column}
\keyword{rename}
