% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votecount_mv.R
\name{votecount_mv}
\alias{votecount_mv}
\title{A function to draw the 'Vote-counting meta-analysis' MetaVolcano}
\usage{
votecount_mv(
  diffexp = list(),
  pcriteria = "pvalue",
  foldchangecol = "Log2FC",
  genenamecol = "Symbol",
  geneidcol = NULL,
  pvalue = 0.05,
  foldchange = 0,
  metathr = 0.01,
  collaps = FALSE,
  jobname = "MetaVolcano",
  outputfolder = ".",
  draw = "HTML"
)
}
\arguments{
\item{diffexp}{list of data.frame/data.table (s) with DE results where lines
are genes}

\item{pcriteria}{the column name of the Pval criteria to consider <string>}

\item{foldchangecol}{the column name of the foldchange variable <string>}

\item{genenamecol}{the column name of the gene name variable <string>}

\item{geneidcol}{the column name of the gene ID/probe/oligo/transcript 
variable <string>}

\item{pvalue}{the Pval to use as threshold c(0:1) <double>}

\item{foldchange}{the foldchange to use as DE threshold c(-Inf: Inf) <double>}

\item{metathr}{the proportion of studies a gene has to be DEG to be 
considered cDEG <double>}

\item{collaps}{if probes should be collapsed based on the DE 
direction <logical>}

\item{jobname}{name of the running job <string>}

\item{outputfolder}{/path where to write the results/}

\item{draw}{wheather or not to draw a .pdf or .html visualization 
<c(NULL, 'PDF', 'HTML')>}
}
\value{
MetaVolcano object
}
\description{
This function draws the vote-counting meta-analysis MetaVolcano
}
\examples{
data(diffexplist)
mv <- votecount_mv(diffexplist)
str(mv)
}
\keyword{'vote-counting}
\keyword{meta-analysis'}
\keyword{metavolcano}
\keyword{write}
